+++ Bonus.md
@@ -3,4 +3,5 @@
 ## List of Features
 
 * Improve UI.
+  * Add cards for questions on main search page.
 * Improve Sitewide color pallete.
