+++ public/js/log-in.js
@@ -1,5 +1,6 @@
 const loginForm = document.querySelector(".log-in-form");
-// console.log(loginForm);
+// const { query } = require("express");
+import { handleErrors } from "./errors.js";
 
 loginForm.addEventListener("submit", async (e) => {
 	e.preventDefault();
@@ -27,34 +28,18 @@ loginForm.addEventListener("submit", async (e) => {
 		localStorage.setItem("COREDUMP_ACCESS_TOKEN", token);
 		localStorage.setItem("COREDUMP_CURRENT_USER_ID", id);
 
-		window.location.href = "/";
+		window.location.href = "http://localhost:8080/";
+
+		document
+			.querySelector(".navbar-logout-text")
+			.classList.remove(".hidden")
+
+		document
+			.querySelector(".navbar-signup-login")
+			.classList.add(".hidden")
+
 	}
 	catch (err) {
-		if (err.status >= 400 && err.status < 600) {
-			const errorJSON = await err.json();
-			const errorsContainer = document.querySelector(".errors-container");
-			let errorsHtml = [
-				`
-			<div class="alert alert-danger">
-				Something went wrong. Please try again.
-			</div>
-		  `,
-			];
-			const { errors } = errorJSON;
-			if (errors && Array.isArray(errors)) {
-				errorsHtml = errors.map(
-					(message) => `
-			  <div class="alert alert-danger">
-				  ${message}
-			  </div>
-			`
-				);
-			}
-			errorsContainer.innerHTML = errorsHtml.join("");
-		} else {
-			alert(
-				"Something went wrong. Please check your internet connection and try again!"
-			);
-		}
+		handleErrors(err);
 	}
 })
