+++ views/login.pug
@@ -35,5 +35,5 @@ html(lang="en")
               input#password.form-control(type='password', name="password", placeholder='Password')
             button.btn.btn-primary(type='submit') Log In
           a(href='/sign-up') Don't have an account? Sign up here.
-        script(src="js/log-in.js")
+        script(src="js/log-in.js" type='module')
     div.gradient
